pyproject.toml
[project]
name = "boko-hacks-2025"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.14.1",
    "blinker==1.9.0",
    "captcha==0.7.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "coverage==7.6.10",
    "Flask-Migrate==4.1.0",
    "Flask-SQLAlchemy==3.1.1",
    "Flask==3.1.0",
    "Flask-WTF==1.2.2",
    "greenlet==3.1.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.5",
    "Mako==1.3.9",
    "MarkupSafe==3.0.2",
    "pillow==11.1.0",
    "pre-commit==4.0.1",
    "pytest==8.3.2",
    "requests==2.32.3",
    "SQLAlchemy==2.0.37",
    "typing_extensions==4.12.2",
    "urllib3==2.3.0",
    "Werkzeug==3.1.3",
    "bleach==6.2.0",
    "python-dotenv==1.0.1"
]

[tool.ruff.format]
# Use `\n` line endings for all files
line-ending = "lf"

[tool.pytest.ini_options]
pythonpath = "."

[tool.coverage.run]
branch = true
relative_files = true

[tool.coverage.report]
show_missing = true
skip_covered = false
include = [
    "lib/**/*.py",
    "boot.py",
    "main.py",
    "repl.py",
    "safemode.py",
    ]
omit = [
    "lib/adafruit_*/**",
    "lib/asyncio_*/**",
    "lib/rv3028*/**",
    "lib/neopixel.py",
    ]

[tool.coverage.html]
directory = ".coverage-reports/html"

[tool.coverage.xml]
output = ".coverage-reports/coverage.xml"
